Update to MSFT_SPBlobCacheSettings.psm1
In Set-TargetResource configuring the max-age attribute in a few farm fails because this attribute is not part of the default Blob Cache attributes and $webconfig.configuration.SharePoint.BlobCache."max-age"  = $changes.MaxAgeInSeconds.ToString() only updates an existing attribute and will not create a new one. The following error message is seen in the console:

PowerShell DSC resource MSFT_SPBlobCacheSettings  failed to execute Set-TargetResource functionality with error
message: Exception setting "max-age": "The property 'max-age' cannot be found on this object. Verify that the property
exists and can be set."
    + CategoryInfo          : InvalidOperation: (:) [], CimException
    + FullyQualifiedErrorId : ProviderOperationExecutionFailure
    + PSComputerName        : localhost

Fix:
Updated the max-age section to use the following, which will create the attribute if it does not exist and update it if it does exist.
$webconfig.configuration.SharePoint.BlobCache.SetAttribute("max-age",$($changes.MaxAgeInSeconds.ToString()))