Update MSFT_SPSearchContentSource.psm1 issue #840
Regarding issue MSFT_SPSearchContentSource does not take StartHour parameter for schedule #840

I was able to solve this issue, but I must admit that I am a newbie regarding the pull request process...

So here it is. From what I was able to understand, the StartHour parameter is correctly checked in the test Get-SPDSCSearchCrawlSchedule.psm1 module.
The issue seems to be in the MSFT_SPSearchContentSource.psm1, as in its Set-TargetResource function there is never the parameter CrawlScheduleStartDateTime which is required by the sharepoint powershell cmdlet Set-SPEnterpriseSearchCrawlContentSource to had a start hour to the schedule.

Hope it helps.